msm: vidc: Fix broken debugfs creation error checks and error paths
The debugfs helper functions do not always return NULL when they fail;
instead, they can return an error number casted as a pointer, so that their
users have the option to determine the exact cause of failure.

Use the IS_ERR_OR_NULL() helper when checking for debugfs errors to fix the
error checks.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: Jago Gardiner (nysadev) <jagogardiner@gmail.com>
Signed-off-by: nysascape <nysadev@raphielgang.org>
Signed-off-by: Yasir-Siddiqui <www.mohammad.yasir.s@gmail.com>
Signed-off-by: Ratoriku <a1063021545@gmail.com>